Add per cast user auth

+++ castpage/views.py
@@ -1,4 +1,5 @@
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
 from django.shortcuts import get_object_or_404, redirect, render
 from .forms import PageSectionForm
 from .models import Cast, PageSection
@@ -12,6 +13,8 @@ def cast_home(request, slug: str):
 @login_required
 def cast_section_new(request, slug: str):
     cast = get_object_or_404(Cast, slug=slug)
+    if not cast.is_manager(request.user.pk):
+        return HttpResponseForbidden()
     if request.method == 'POST':
         form = PageSectionForm(request.POST)
         if form.is_valid():
@@ -26,8 +29,9 @@ def cast_section_new(request, slug: str):
 @login_required
 def cast_section_edit(request, slug: str, pk: int):
     cast = get_object_or_404(Cast, slug=slug)
+    if not cast.is_manager(request.user.pk):
+        return HttpResponseForbidden()
     section = get_object_or_404(PageSection, pk=pk)
-    print(section)
     if request.method == 'POST':
         form = PageSectionForm(request.POST, instance=section)
         if form.is_valid():
@@ -42,6 +46,8 @@ def cast_section_edit(request, slug: str, pk: int):
 @login_required
 def cast_section_remove(request, slug: str, pk: int):
     cast = get_object_or_404(Cast, slug=slug)
+    if not cast.is_manager(request.user.pk):
+        return HttpResponseForbidden()
     section = get_object_or_404(PageSection, pk=pk)
     section.delete()
     return redirect('cast_home', slug=cast.slug)
